working on routing

+++ db/migrate/20191120211645_create_users.rb
@@ -7,7 +7,7 @@ class CreateUsers < ActiveRecord::Migration[5.2]
       t.string :password_digest
       t.string :img_url, default: "http://brandesscadmusrealestate.com/images/team/blank.png"
       t.string :bg_url, default: "https://cdn.hipwallpaper.com/i/39/64/bzVo03.jpg"
-      t.integer :coinbank, default:  1000
+      t.bigint :coinbank, default:  1000
       
 
       t.timestamps
