+++ db/schema.rb
@@ -49,7 +49,7 @@ ActiveRecord::Schema.define(version: 2019_12_27_150827) do
     t.string "password_digest"
     t.string "img_url", default: "http://brandesscadmusrealestate.com/images/team/blank.png"
     t.string "bg_url", default: "https://cdn.hipwallpaper.com/i/39/64/bzVo03.jpg"
-    t.integer "coinbank", default: 1000
+    t.bigint "coinbank", default: 1000
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
